Do coding for 3a
Largely this was rearranging code for notebackend, but also needed a 
small tweak to the path for api/notes -- whether to start with a '/' or 
not. Could probably have also changed this in the notebackend to jive 
with what the frontend was sending, and got it to work also -- opted to 
change the frontend, though.
