+++ osa2/notes/src/services/notes.js
@@ -1,7 +1,7 @@
 import axios from 'axios'
 // const baseUrl = 'http://localhost:3001/api/notes'
 // const baseUrl = 'https://ancient-ocean-58745.herokuapp.com/notes'
-const baseUrl = 'api/notes'
+const baseUrl = '/api/notes'
 
 const getAll = () => {
     return axios
