+++ osa3/notebackend/models/note.js
@@ -1,21 +1,5 @@
 const mongoose = require('mongoose');
 
-// Make Mongoose use `findOneAndUpdate()`. Note that this option is `true`
-// by default, you need to set it to false.
-mongoose.set('useFindAndModify', false);
-
-const url = process.env.MONGODB_URI
-
-console.log('connecting to: ', url);
-
-mongoose.connect(url, { useNewUrlParser: true, useUnifiedTopology: true })
-  .then(() => {
-    console.log('connected to MongoDB');
-  })
-  .catch((error) => {
-    console.log('error connecting to MongoDB, error: ', error.message);
-  });
-
 const noteSchema = new mongoose.Schema({
   content: {
     type: String,
